Switch hashes to uint64, fix example
The example now hashes the tasks in a nicer way, using actual hashing
techniques.

I've also switched the hash key over to uint64.

+++ pkg/nbt/nbt.go
@@ -2,7 +2,7 @@ package nbt
 
 type Task interface {
 	/* Returns a hash for this task that can be used to use it in a map. */
-	Hash() uintptr
+	Hash() uint64
 	/* Returns true if this task matches the given task, false otherwise. */
 	Matches(Task) bool
 	Perform(h Handler) error
